updating to kotlin 1.5.0

+++ kool-core/src/commonMain/kotlin/de/fabmax/kool/modules/audio/WavFile.kt
@@ -127,7 +127,7 @@ class WavFile(fileData: Uint8Buffer) {
     private fun ByteArray.toCharArray(): CharArray {
         val arr = CharArray(size)
         for (i in indices) {
-            arr[i] = this[i].toChar()
+            arr[i] = this[i].toInt().toChar()
         }
         return arr
     }
